Added monitoring capabilities via JMX
A servlet was written that can be integrated into Se Grid during startup.
This servlet when initialized will start JMX and will expose data about
Se Grid overall state, it's possible e.g. to get a notion of how it's loaded.
Later on this functionality mighbe extracted and a pull request can be sent
to Selenium developers.

Readme files were splitted for each module (tests and monitoring) so that main
Readme doesn't get blottered with a lot of information.